create common utils across project
