+++ src/common-utils/index.js
@@ -1,0 +1,56 @@
+'use strict';
+const { makeRequest } = require('../request');
+
+function base64Encode(encodeData) {
+    const buff = Buffer.from(encodeData);
+    return buff.toString('base64');
+}
+let headers = {};
+module.exports = {
+    setAccessToken: function (token) {
+        this.options.access_token = token;
+    },
+    getAccessToken: function () {
+        if (!this.options.clientID) throw new Error('Missing Client ID');
+        if (!this.options.clientSecret) throw new Error('Missing Client Secret or Cert Id');
+        if (!this.options.body) throw new Error('Missing Body, required Grant type');
+        const encodedStr = base64Encode(this.options.clientID + ':' + this.options.clientSecret);
+        const self = this;
+        const auth = 'Basic ' + encodedStr;
+        return makeRequest(this.options.baseUrl, '/identity/v1/oauth2/token', 'POST', this.options.body, auth).then((result) => {
+            const resultJSON = JSON.parse(result);
+            self.setAccessToken(resultJSON.access_token);
+            return resultJSON;
+        });
+    },
+    setHeaders(headerObj) {
+        headers = { ...headers, ...headerObj };
+    },
+    getHeaders() {
+        return headers;
+    },
+    upperCase(data) {
+        if (!isString(data)) data = data.toString();
+        return data.toUpperCase();
+    },
+
+    // Returns if a value is a string
+    isString(value) {
+        return typeof value === 'string' || value instanceof String;
+    },
+
+    // Returns if object is empty or not
+    isEmptyObj(obj) {
+        for (let key in obj) {
+            if (obj.hasOwnProperty(key)) {
+                return false;
+            }
+        }
+        return true;
+    },
+
+    encodeURLQuery(url) {
+        return encodeURIComponent(url).replace(/'/g, '%27').replace(/"/g, '%22');
+    },
+
+};
