+++ src/index.js
@@ -1,5 +1,4 @@
 'use strict';
-const { makeRequest, base64Encode } = require('./request');
 const { getItem,
     getItemByLegacyId,
     getItemByItemGroup,
@@ -17,13 +16,29 @@ const { findItemsByKeywords,
     findItemsByCategory,
     findCompletedItems,
     getVersion } = require('./findingApi');
+const { setAccessToken,
+    getAccessToken,
+    setHeaders,
+    getHeaders,
+    isEmptyObj } = require('./common-utils');
 const { getSimilarItems, getMostWatchedItems } = require('./merchandising');
 const { PROD_BASE_URL, SANDBOX_BASE_URL, BASE_SANDBX_SVC_URL, BASE_SVC_URL } = require('./constants');
 const PROD_ENV = 'PROD';
 const SANDBOX_ENV = 'SANDBOX';
 
-function Ebay(options) {
 
+/**
+ * Creates a eBay instance.
+ *
+ * @param {Object} options configuration options
+ * @param {String} options.clientID Client Id/App id
+ * @param {String} options.env Environment, defaults to PROD
+ * @param {String} options.headers HTTP request headers
+ * @constructor
+ * @public
+ */
+
+function Ebay(options) {
     if (!options) throw new Error('Options is missing, please provide the input');
     if (!options.clientID) throw Error('Client ID is Missing\ncheck documentation to get Client ID http://developer.ebay.com/DevZone/account/');
     if (!(this instanceof Ebay)) return new Ebay(options);
@@ -40,24 +55,10 @@ function Ebay(options) {
 }
 
 Ebay.prototype = {
-
-    setAccessToken: function (token) {
-        this.options.access_token = token;
-    },
-
-    getAccessToken: function () {
-        if (!this.options.clientID) throw new Error('Missing Client ID');
-        if (!this.options.clientSecret) throw new Error('Missing Client Secret or Cert Id');
-        if (!this.options.body) throw new Error('Missing Body, required Grant type');
-        const encodedStr = base64Encode(this.options.clientID + ':' + this.options.clientSecret);
-        const self = this;
-        const auth = 'Basic ' + encodedStr;
-        return makeRequest(this.options.baseUrl, '/identity/v1/oauth2/token', 'POST', this.options.body, auth).then((result) => {
-            const resultJSON = JSON.parse(result);
-            self.setAccessToken(resultJSON.access_token);
-            return resultJSON;
-        });
-    },
+    setAccessToken,
+    getAccessToken,
+    setHeaders,
+    getHeaders,
     findItemsByKeywords,
     findItemsByCategory,
     findCompletedItems,
